+++ lib/src/widgets/file_bubble.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:v_platform/v_platform.dart';
+
 import '../core/constants.dart';
 import '../core/enums.dart';
 import '../core/models.dart';
@@ -90,6 +91,7 @@ class VFileBubble extends BaseBubble {
                       ),
                       BubbleSpacing.vGapXS,
                       Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
                           Flexible(
                             child: Row(